refactor : 객체 상태를 내장 타입에게 책임 할당하도록 변경
