+++ src/main/java/com/yapp/archiveServer/archive/domain/artwork/Artwork.java
@@ -1,7 +1,10 @@
-package com.yapp.archiveServer.archive.domain;
+package com.yapp.archiveServer.archive.domain.artwork;
 
+import com.yapp.archiveServer.archive.domain.exhibit.Exhibit;
 import com.yapp.archiveServer.global.common.BaseEntity;
-import lombok.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.hibernate.annotations.DynamicInsert;
 
 import javax.persistence.*;
@@ -12,6 +15,7 @@ import javax.persistence.*;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Artwork extends BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -20,21 +24,9 @@ public class Artwork extends BaseEntity {
     @JoinColumn(name = "post_id", nullable = false)
     private Exhibit exhibit;
 
-    @Column(nullable = false)
-    private String url;
+    @Embedded
+    private ExposureCondition condition;
 
-    @Column(nullable = false)
-    private int seq;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String artist;
-
-    @Column(nullable = false, columnDefinition = "boolean default false")
-    private boolean isMainImage;
-
-    @Column(columnDefinition = "text")
-    private String description;
+    @Embedded
+    private ArtworkContents contents;
 }
